+++ info/2024_qualifiers.yml
@@ -1,0 +1,68 @@
+- Purdue
+- Green Bay
+- Southern
+- New Mexico
+- St. John's
+- Baylor
+- Appalachian State
+- Auburn
+- Akron
+- Utah State
+- Grand Canyon
+- Kansas
+- Drexel
+- Clemson
+- Nebraska
+- Marquette
+- Colgate
+- North Carolina
+- Central Connecticut
+- North Carolina Central
+- TCU
+- Mississippi State
+- BYU
+- McNeese State
+- Illinois
+- Samford
+- South Carolina
+- Indiana State
+- Creighton
+- Morehead State
+- Texas Tech
+- Saint Mary's
+- Arizona
+- Eastern Washington
+- UConn
+- Eastern Kentucky
+- Michigan State
+- Texas
+- Kentucky
+- Richmond
+- Iowa State
+- Cornell
+- San Diego State
+- Seton Hall
+- Memphis
+- Duke
+- Vermont
+- Northwestern
+- Providence
+- Tennessee
+- Quinnipiac
+- Houston
+- So. Dakota State
+- Ole Miss
+- Utah
+- Florida Atlantic
+- UC-Irvine
+- Dayton
+- Charlotte
+- Oklahoma
+- Boise State
+- Colorado
+- Alabama
+- High Point
+- Colorado State
+- Texas A&M
+- Wisconsin
+- Sam Houston State
