--- shop_config.yml
@@ -1,28 +1,0 @@
-# this is stuff for posting to shop
-base_url : "https://api.printify.com/v1"
-cfb:
-  shop_id : 13947493
-  shop_name: '70f8b8-5'
-rand:
-  shop_id : 13947121
-  shop_name : '56dc74-2'
-cbb:
-  shop_id : 14069125
-  shop_name : c2b32f-2
-title: 'test'
-price: 2599
-description: 'this a test'
-tags: 'this,a,test'
-blueprint_id: 6
-print_provider_id: 99
-variant_ids:
-  - 12070
-  - 12071
-  - 12072
-  - 12073
-  - 12074
-  - 12100
-  - 12101
-  - 12102
-  - 12103
-  - 12104
